Refactor - flytter på sakquery som brukes felles ut i common
