+++ frontend/src/common/sak/useStonadsTypeNavn.ts
@@ -1,13 +1,12 @@
-import { useSuspenseQuery } from "@tanstack/react-query";
-import { getKodeverkStonadsTypeOptions } from "~/routes/sak/$saksnummer/-api/sak.query";
 import type { StonadType } from "~/common/sak/sak.types";
+import { useStonadsType } from "~/common/sak/useStonadsType";
 
 export function useStonadsTypeNavn() {
-    const { data: stonadsTyper } = useSuspenseQuery(getKodeverkStonadsTypeOptions());
+    const stonadsType = useStonadsType();
     return (type?: StonadType) => {
         if (!type) {
             return "";
         }
-        return stonadsTyper.find((t) => t.type === type)?.navn ?? type;
+        return stonadsType(type)?.navn ?? type;
     };
 }
