+++ frontend/src/routes/oppgave/$oppgaveid/-components/StonadsTypeVelger.tsx
@@ -1,7 +1,7 @@
-import {Select} from "@navikt/ds-react";
-import {useSuspenseQuery} from "@tanstack/react-query";
-import {getKodeverkStonadsTypeOptions} from "~/routes/sak/$saksnummer/-api/sak.query";
-import type {StonadType} from "~/common/sak/sak.types";
+import { Select } from "@navikt/ds-react";
+import { useSuspenseQuery } from "@tanstack/react-query";
+import { getKodeverkStonadsTypeOptions } from "~/common/sak/sak.query";
+import type { StonadType } from "~/common/sak/sak.types";
 
 interface Props {
     value?: StonadType;
