+++ frontend/src/routes/sak/$saksnummer/-api/useInvalidateSakQuery.ts
@@ -1,7 +1,7 @@
 import { hentEndringsloggForSakQueryKey } from "@generated/@tanstack/react-query.gen";
 import { useQueryClient } from "@tanstack/react-query";
-import { apiFinnJournalpostForSakQueryKey } from "~/routes/sak/$saksnummer/-api/journalpost.query";
-import { sakQueryKey } from "~/routes/sak/$saksnummer/-api/sak.query";
+import { sakQueryKey } from "~/common/sak/sak.query";
+import { apiFinnJournalpostForSakQueryKey } from "./journalpost.query";
 
 /** Invaliderer cachen for en sak og tilhørende endringslogg */
 export function useInvalidateSakQuery() {
