+++ frontend/src/routes/sak/$saksnummer/-components/Fritekstbrev.tsx
@@ -1,11 +1,11 @@
-import {sendBrevMutation} from "@generated/@tanstack/react-query.gen";
-import {Dialog, ErrorSummary} from "@navikt/ds-react";
-import {useMutation, useQuery, useSuspenseQuery} from "@tanstack/react-query";
-import {useParams} from "@tanstack/react-router";
-import {getOrCreateBrevOptions} from "~/routes/sak/$saksnummer/-api/brev.query";
-import {getSakOptions} from "~/routes/sak/$saksnummer/-api/sak.query";
-import {useInvalidateSakQuery} from "~/routes/sak/$saksnummer/-api/useInvalidateSakQuery";
-import {BrevEditor} from "~/routes/sak/$saksnummer/-components/BrevEditor";
+import { sendBrevMutation } from "@generated/@tanstack/react-query.gen";
+import { Dialog, ErrorSummary } from "@navikt/ds-react";
+import { useMutation, useQuery, useSuspenseQuery } from "@tanstack/react-query";
+import { useParams } from "@tanstack/react-router";
+import { getSakOptions } from "~/common/sak/sak.query";
+import { getOrCreateBrevOptions } from "~/routes/sak/$saksnummer/-api/brev.query";
+import { BrevEditor } from "~/routes/sak/$saksnummer/-components/BrevEditor";
+import { useInvalidateSakQuery } from "../-api/useInvalidateSakQuery";
 
 interface FritekstBrevProps {
     open: boolean;
@@ -15,8 +15,7 @@ interface FritekstBrevProps {
 export function FritekstBrev({ open, onOpenChange }: FritekstBrevProps) {
     const { saksnummer } = useParams({ from: "/sak/$saksnummer" });
     const { data: sak } = useSuspenseQuery(getSakOptions(saksnummer));
-    const kanSendeBrev =
-        sak.rettigheter.includes("FRITEKSTBREV") ;
+    const kanSendeBrev = sak.rettigheter.includes("FRITEKSTBREV");
     const { data: brev } = useQuery({
         ...getOrCreateBrevOptions(saksnummer, "FRITEKSTBREV", "BRUKER"),
         enabled: open,
