+++ frontend/src/routes/sak/$saksnummer/route.tsx
@@ -11,6 +11,7 @@ import { finnPersonQuery } from "~/common/person/person.query";
 import { ProcessMenu } from "~/common/process-menu/ProcessMenu";
 import { StepType } from "~/common/process-menu/StepType";
 import { useSakshistorikkAntall } from "~/common/sak/historikk/useSakshistorikkAntall";
+import { getSakOptions } from "~/common/sak/sak.query";
 import type { TilstandStatusType } from "~/common/sak/sak.types";
 import { isSakFerdig } from "~/common/sak/sak.utils";
 import { kortNavn, kortSaksnummer } from "~/common/string.utils";
@@ -20,7 +21,6 @@ import SakAlert from "~/routes/sak/$saksnummer/-components/SakAlerts";
 import SakEndringer from "~/routes/sak/$saksnummer/-components/SakEndringer";
 import SakOppsummering from "~/routes/sak/$saksnummer/-components/SakOppsummering";
 import { SakshistorikkSakTabell } from "~/routes/sak/$saksnummer/-components/SakshistorikkSakTabell";
-import { getSakOptions } from "./-api/sak.query";
 import BehandlingsMeny from "./-components/BehandlingsMeny";
 
 export const Route = createFileRoute("/sak/$saksnummer")({
