+++ frontend/src/routes/sak/$saksnummer/sendklage.tsx
@@ -19,9 +19,9 @@ import { useMutation, useQuery, useSuspenseQuery } from "@tanstack/react-query";
 import { createFileRoute, useNavigate } from "@tanstack/react-router";
 import { useState } from "react";
 import { dateTilIsoDato, isoTilLokal } from "~/common/dato.utils";
+import { getSakOptions } from "~/common/sak/sak.query";
 import { useStonadsTypeNavn } from "~/common/sak/useStonadsTypeNavn";
-import { getSakOptions } from "~/routes/sak/$saksnummer/-api/sak.query";
-import { useInvalidateSakQuery } from "~/routes/sak/$saksnummer/-api/useInvalidateSakQuery";
+import { useInvalidateSakQuery } from "./-api/useInvalidateSakQuery";
 
 export const Route = createFileRoute("/sak/$saksnummer/sendklage")({
     component: SendKlagePage,
